Add files via upload
BookSmart is a smart library system that uses IoT and Firebase to manage book and seat availability. It integrates RFID and PIR sensors with an ESP32 to update real-time data to a web interface. Users can log in to reserve seats, and librarians can monitor usage through a dashboard. The system uses HTML, CSS, JavaScript, and Firebase services for authentication and database.